move babel plugin in sepearte file
