+++ packages/next-yak/package.json
@@ -6,11 +6,13 @@
   "types": "dist/index.d.ts",
   "exports": {
     ".": "./dist/index.js",
-    "./withYak": "./loaders/withYak.cjs"
+    "./withYak": "./loaders/withYak.cjs",
+    "./babel": "./runtime/babel-yak-plugin.js"
   },
   "typesVersions": {
     "*": {
       "withYak": ["./loaders/withYak.d.ts"],
+      "babel": ["./runtime/babel-yak-plugin.d.ts"],
       "*": ["./runtime/index.d.ts"]
     }
   },
